+++ BepInExPlugins/LoadoutLockdown/CHANGELOG.md
@@ -1,7 +1,8 @@
 # 1.1.0
 - Moving an equipped weapon out of a weapon slot will unequip it. (unless its configured to not require a hotbar slot)
-- Added `ApplyPvpMenuSwapRulesToPVE` option. Can be enabled to make PVP menu-swapping rules apply to PVE combat too.
+- Added `ApplyPvpMenuSwapRulesToPVE` option for rulesets. Can be enabled to make PVP menu-swapping rules apply to PVE combat too.
 - Fixed an issue where trying to equip a weapon from the inventory would move it into a weapon slot, but not equip it.
+- Added `LogRulings` option to main config. When enabled, logs details about every ruling made.
 
 # 1.0.0
 - Initial release