+++ BepInExPlugins/LoadoutLockdown/CHANGELOG.md
@@ -1,3 +1,6 @@
+# upcoming release
+- Bugfix: Weapons couldn't be unequipped in PVP combat.
+
 # 1.1.0
 - Moving an equipped weapon out of a weapon slot will unequip it. (unless its configured to not require a hotbar slot)
 - Added `ApplyPvpMenuSwapRulesToPVE` option for rulesets. Can be enabled to make PVP menu-swapping rules apply to PVE combat too.